+++ R/asso_test.R
@@ -1,10 +1,9 @@
 #' Perform association tests between phenotype and covariates
 #'
 #' @param pheno phenotype dataframe - output from makePhenotypes function
-#' @param covariates covariate dataframe - output from risk.factor function.
+#' @param covariates covariate dataframe - output from risk_factor function.
 #' @param phe.name Phenotype name in the data.
 #' @param cov.name Selected covariate names in the data. By default, cov.name=c("sex","age","bmi"), covariates include sex, age and BMI.
-#' @param asso.output Name of association test result file to be outputted. By default, asso.output=NULL, it returns results but doesn't generate any files.
 #' @export log_cov
 #' @return Outputs association test results with OR, 95% CI, and p-value.
 #' @import questionr
@@ -15,7 +14,7 @@
 #' log_cov(pheno=phe, covariates=covar, phe.name="hospitalisation", cov.name=c("sex","age","bmi"))
 #' }
 
-log_cov <- function(pheno, covariates, phe.name, cov.name = c("sex","age","bmi"), asso.output=NULL){
+log_cov <- function(pheno, covariates, phe.name, cov.name = c("sex","age","bmi")){
   data <- inner_join(pheno, covariates, by="ID")
   y <- data[,c(phe.name,cov.name)]
   y <- na.omit(y)
@@ -25,6 +24,6 @@ log_cov <- function(pheno, covariates, phe.name, cov.name = c("sex","age","bmi")
   OR <- odds.ratio(m)
   asso <- as.data.frame(cbind(log.reg$coefficients[,1],OR[,c(1:4)]))
   colnames(asso)[1] <- "Estimate"
-  if(!(is.null(asso.output))) write.csv(asso,paste0(asso.output,".csv"),quote = F)
-  return(asso)
+  class(asso) <- "data.frame"
+  asso
 }